Added small crops in DNN detector for High resolution videos and small objects (#75)
* Added hough3d sources

* Calculate 3D hough trajectory lines

* Added fps

* Some warnings removing

* Fixed warning

* Added small crops in DNN detector for High resolution videos and small objects

* Small refactoring

* Some fix for single tracking

* Added openmp flags to the compiler

* Draw trajectory option

* Update Readme

* Fixed link

* Hough3D disabled now in CMake

+++ nms.h
@@ -48,8 +48,8 @@ inline void nms(
             // grab the current rectangle
             const cv::Rect& rect2 = srcRects[pos->second];
 
-            float intArea = (rect1 & rect2).area();
-            float unionArea = rect1.area() + rect2.area() - intArea;
+			float intArea = static_cast<float>((rect1 & rect2).area());
+			float unionArea = static_cast<float>(rect1.area() + rect2.area() - intArea);
             float overlap = intArea / unionArea;
 
             // if there is sufficient overlap, suppress the current bounding box
@@ -122,8 +122,8 @@ inline void nms2(
             // grab the current rectangle
             const cv::Rect& rect2 = srcRects[pos->second];
 
-            float intArea = (rect1 & rect2).area();
-            float unionArea = rect1.area() + rect2.area() - intArea;
+			float intArea = static_cast<float>((rect1 & rect2).area());
+			float unionArea = static_cast<float>(rect1.area() + rect2.area() - intArea);
             float overlap = intArea / unionArea;
 
             // if there is sufficient overlap, suppress the current bounding box
@@ -199,8 +199,8 @@ inline void nms3(
             // grab the current rectangle
             const cv::Rect& rect2 = GetRect(srcRects[pos->second]);
 
-            float intArea = (rect1 & rect2).area();
-            float unionArea = rect1.area() + rect2.area() - intArea;
+			float intArea = static_cast<float>((rect1 & rect2).area());
+			float unionArea = static_cast<float>(rect1.area() + rect2.area() - intArea);
             float overlap = intArea / unionArea;
 
             // if there is sufficient overlap, suppress the current bounding box
